requirements.txt
anthropic==0.49.0
streamlit==1.43.2
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
python-dotenv==1.0.1
openai==1.59.6